- return something in SimulationRuntime/c/util/integer_array.c - hopefully fix condition in SimulationRuntime/c/simulation/solver/initialization/initialization.c - try to use proper data types in function interfaces - get rid of some of the warnings - transform to char* where needed
git-svn-id: https://openmodelica.org/svn/OpenModelica/trunk@23747 f25d12d1-65f4-0310-ae8a-bbce733d8d8e

+++ SimulationRuntime/c/meta/meta_modelica_builtin.c
@@ -355,7 +355,7 @@ modelica_metatype boxptr_stringGetStringChar(threadData_t *threadData,metamodeli
   MMC_CHECK_STRING(str);
   if (ix < 1 || ix > (long) MMC_STRLEN(str))
     MMC_THROW_INTERNAL();
-  return mmc_strings_len1[MMC_STRINGDATA(str)[ix-1]];
+  return mmc_strings_len1[(size_t)MMC_STRINGDATA(str)[ix-1]];
 }
 
 modelica_integer nobox_stringGet(threadData_t *threadData,metamodelica_string str, modelica_integer ix)
